move tests into testPerf
