+++ packages/core/src/testInteg/shared/utilities/workspaceUtils.test.ts
@@ -18,8 +18,6 @@ import globals from '../../../shared/extensionGlobals'
 import { CodelensRootRegistry } from '../../../shared/fs/codelensRootRegistry'
 import { createTestWorkspace, createTestWorkspaceFolder, toFile } from '../../../test/testUtil'
 import sinon from 'sinon'
-import { performanceTest } from '../../../shared/performance/performance'
-import { randomUUID } from '../../../shared/crypto'
 import { fs } from '../../../shared'
 
 describe('findParentProjectFile', async function () {
@@ -326,56 +324,6 @@ describe('collectFiles', function () {
         assert.deepStrictEqual(1, result.length)
         assert.deepStrictEqual('non-license.md', result[0].relativeFilePath)
     })
-
-    performanceTest(
-        // collecting all files in the workspace and zipping them is pretty resource intensive
-        {
-            linux: {
-                userCpuUsage: 85,
-                heapTotal: 2,
-                duration: 0.8,
-            },
-        },
-        'calculate cpu and memory usage',
-        function () {
-            const totalFiles = 100
-            return {
-                setup: async () => {
-                    const workspace = await createTestWorkspaceFolder()
-
-                    sinon.stub(vscode.workspace, 'workspaceFolders').value([workspace])
-
-                    const fileContent = randomUUID()
-                    for (let x = 0; x < totalFiles; x++) {
-                        await toFile(fileContent, path.join(workspace.uri.fsPath, `file.${x}`))
-                    }
-
-                    return {
-                        workspace,
-                    }
-                },
-                execute: async ({ workspace }: { workspace: vscode.WorkspaceFolder }) => {
-                    return {
-                        result: await collectFiles([workspace.uri.fsPath], [workspace], true),
-                    }
-                },
-                verify: (
-                    _: { workspace: vscode.WorkspaceFolder },
-                    { result }: { result: Awaited<ReturnType<typeof collectFiles>> }
-                ) => {
-                    assert.deepStrictEqual(result.length, totalFiles)
-                    const sortedFiles = [...result].sort((a, b) => {
-                        const numA = parseInt(a.relativeFilePath.split('.')[1])
-                        const numB = parseInt(b.relativeFilePath.split('.')[1])
-                        return numA - numB
-                    })
-                    for (let x = 0; x < totalFiles; x++) {
-                        assert.deepStrictEqual(sortedFiles[x].relativeFilePath, `file.${x}`)
-                    }
-                },
-            }
-        }
-    )
 })
 
 describe('getWorkspaceFoldersByPrefixes', function () {
