+++ packages/core/src/testInteg/testPerf/tryInstallLsp.test.ts
@@ -7,10 +7,10 @@ import sinon from 'sinon'
 import { Content } from 'aws-sdk/clients/codecommit'
 import AdmZip from 'adm-zip'
 import path from 'path'
-import { LspController } from '../amazonq'
-import { fs, getRandomString, globals } from '../shared'
-import { createTestWorkspace } from '../test/testUtil'
-import { performanceTest } from '../shared/performance/performance'
+import { LspController } from '../../amazonq'
+import { fs, getRandomString, globals } from '../../shared'
+import { createTestWorkspace } from '../../test/testUtil'
+import { performanceTest } from '../../shared/performance/performance'
 
 // fakeFileContent is matched to fakeQServerContent based on hash.
 const fakeHash = '4eb2865c8f40a322aa04e17d8d83bdaa605d6f1cb363af615240a5442a010e0aef66e21bcf4c88f20fabff06efe8a214'
